+++ docs/13_0_0/gettingstarted/whatsnew.md
@@ -6,6 +6,8 @@ This page contains a list of big features. Please check the GitHub issues for al
 
 ### PrimeFaces
 
+* BlockUI
+    * Added attribute `delay` to delay displaying similar to AjaxStatus.
 * Captcha
     * Added attribute `sourceUrl` to override the Google JS location for countries that do not have access to Google.
 * DataGrid